image_caption/requirements.txt
tensorflow-gpu==2.10.0
keras==2.10.0
protobuf==3.19.6
numpy==1.24.4
matplotlib==3.7.5
scikit-learn==1.3.2
pandas==2.2.2
Pillow==10.3.0
tqdm==4.66.4
streamlit
tensorflow-hub==0.12.0
opencv-python